Animate the order total to make it more exciting
